+++ tortoise/migrations/schema_generator/state_diff.py
@@ -115,7 +115,6 @@ class StateModelDiff:
             return []
 
         operations: list[TortoiseOperation] = []
-        operations.extend(StateFieldDiff(self.old_state, self.new_state).generate_operations())
         operations.extend(self._generate_index_operations())
         operations.extend(self._generate_constraint_operations())
         old_options = _model_options_for_compare(self.old_state.options)
@@ -128,6 +127,12 @@ class StateModelDiff:
                 )
             )
 
+        operations.extend(StateFieldDiff(self.old_state, self.new_state).generate_operations())
+
+        # Classes that must always be last
+        always_last = (AddIndex, AddConstraint)
+
+        operations = sorted(operations, key=lambda op: isinstance(op, always_last))
         return operations
 
     @staticmethod