+++ tortoise/migrations/schema_generator/state_diff.py
@@ -128,6 +128,10 @@ class StateModelDiff:
             )
 
         operations.extend(StateFieldDiff(self.old_state, self.new_state).generate_operations())
+
+        # Classes that must always be last
+        always_last = (AddIndex, AddConstraint)
+        operations = sorted(operations, key=lambda op: isinstance(op, always_last))
         return operations
 
     @staticmethod